requirements.txt
# A2A Bridge to uAgent Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
pydantic>=2.10.6
python-dotenv>=1.1.0
httpx>=0.28.1
click>=8.1.8

# Web server
uvicorn>=0.30.1,<0.31.0

# LangChain and AI dependencies
langchain-google-genai>=2.0.10
langgraph>=0.3.18
langchain-openai>=0.1.0
langchain-core

# uAgent framework
uagents==0.22.5
uagents-ai-engine>=0.4.0

# A2A SDK
a2a-sdk==0.2.8

# Additional dependencies for specific functionality
requests>=2.31.0
fastapi>=0.104.0
starlette>=0.37.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment management
python-decouple>=3.8
